# apiDados/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api Hypatiamat",
        "description": "Aqui encontra-se a documentação de toda a api de dados do Hypatiamat, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3050",
    "basePath": "/",
...
    "paths": {
...
        "/aplicacoes/alunos/{idAluno}/escola":{
          "put":{
            "tags": [
              "Alunos"
            ],
            "summary": "Dado um identificador de um aluno, edite a escola dele.",
            "description": "Editar a escola de um aluno.",
            "operationId": "putEscolaAlunoById",
            "parameters": [
              {
                "name": "idAluno",
                "in": "path",
                "description": "Identificador de um aluno.",
                "required": true,
                "type": "integer"
              },
              {
                "name": "escola",
                "in": "formData",
                "description": "Nova escola do aluno (identificador da escola)",
                "required": true,
                "type": "integer"
              }
            ],
            "responses": {
              "200": {
                "description": "A escola do aluno foi alterada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
              "security": [
                {
                  "api_key": []
                }
              ]
          }
        },
...
        "/aplicacoes/escolas": {
          "get": {
              "tags":[
                  "Escolas"
              ],
              "summary": "Todos as escolas registados.",
              "description": "Retorna uma lista com a informação à cerca das escolas.",
              "responses": {
                  "200": {
                      "description": "Lista de Escolas",
                      "schema": {
                        "type":"array",
                        "items":{
                          "$ref": "#/definitions/Escola"
                        }
                      }
                  },
                  "401":{
                    "description": "Falta-lhe o token de acesso ou o token não é válido."
                  }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
          },
          "post":{
              "tags":[
                  "Escolas"
              ],
              "summary": "Inserir uma nova escola.",
              "description": "Adiciona uma nova escola à base de dados.",
              "parameters": [
                  {
                    "name": "nome",
                    "in": "formData",
                    "description": "Nome da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "localidade",
                    "in": "formData",
                    "description": "Localidade da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "distrito",
                    "in": "formData",
                    "description": "Distrito da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "pais",
                    "in": "formData",
                    "description": "País da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "cod",
                    "in": "formData",
                    "description": "Código da escola",
                    "required": true,
                    "type": "string"
                  }
                ],
              "responses": {
                  "200": {
                      "description": "Retorna o id da escola criada",
                      "schema": {
                          "type": "string",
                          "items": {
                              "properties": {
                                  "id": {
                                      "type": "string"
                                  }
                              }
                          }
                      }
                  }
              }
          }
      },
      "/aplicacoes/escolas/{idEscola}":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um identifcador da escola, devolve a informação relativa a essa escola.",
          "description": "Devolve a informação relativa a uma escola.",
          "operationId": "getEscolaById",
          "parameters": [
            {
              "name": "idEscola",
              "in": "path",
              "description": "Identificador da escola a retornar.",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna a informação relativa a uma escola.",
                "schema": {
                    "$ref": "#/definitions/Escola"   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        },
        "delete":{
            "tags": [
              "Escolas"
            ],
            "summary": "Dado um identifcador da escola, pode apagá-la.",
            "description": "Apagar a informação da escola.",
            "operationId": "deleteEscolaById",
            "parameters": [
              {
                "name": "idEscola",
                "in": "path",
                "description": "Identificador da escola a apagar.",
                "required": true,
                "type": "integer"
              }
            ],
            "responses": {
              "200": {
                  "description": "A escola foi apagada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
      },
      "/aplicacoes/escolas/{idEscola}/alunos":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um identifcador da escola, devolve todos os alunos da escola.",
          "description": "Devolve todos os alunos de uma escola.",
          "operationId": "getAlunosByEscola",
          "parameters": [
            {
              "name": "idEscola",
              "in": "path",
              "description": "Identificador da escola.",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todos os alunos da escola.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Aluno"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/aplicacoes/escolas/{idEscola}/professores":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um identifcador da escola, devolve todos os professores da escola.",
          "description": "Devolve todos os professores de uma escola.",
          "operationId": "getProfessoresByEscola",
          "parameters": [
            {
              "name": "idEscola",
              "in": "path",
              "description": "Identificador da escola.",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todos os professores da escola.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Professor"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/aplicacoes/escolas/paises/{pais}":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um país, devolve todos escolas desse país.",
          "description": "Devolve todos as escolas de um país.",
          "operationId": "getEscolasbyPais",
          "parameters": [
            {
              "name": "pais",
              "in": "path",
              "description": "Pais das escolas.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todos as escolas do país.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Escola"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/aplicacoes/escolas/localidades/{localidade}":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dada uma localidade, devolve todas escolas dessa localidade.",
          "description": "Devolve todos as escolas de uma localidade.",
          "operationId": "getEscolasbyLocalidade",
          "parameters": [
            {
              "name": "localidade",
              "in": "path",
              "description": "Localidade das escolas.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todas as escolas da localidade.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Escola"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/aplicacoes/escolas/distritos/{distrito}":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um distrito, devolve todos escolas desse distrito.",
          "description": "Devolve todos as escolas de um distrito.",
          "operationId": "getEscolasbyDistrito",
          "parameters": [
            {
              "name": "distrito",
              "in": "path",
              "description": "Distrito das escolas.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todos as escolas do distrito.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Escola"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
...
    },
    "securityDefinitions":{
      "api_key":{
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "security": [{
       "api_key": []
    }],
    "definitions": {
        "Aluno": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "user": {
                "type": "string"
              },
              "numero": {
                "type": "integer"
              },
              "nome": {
                "type": "string"
              },
              "datanascimento": {
                "type": "string"
              },
              "escola": {
                "type": "string"
              },
              "turma":{
                "type": "string"
              },
              "email":{
                "type":"string"
              },
              "confirmacao":{
                "type":"integer"
              }
            }
        },
        "Professor":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "codigo": {
              "type": "string"
            },
            "nome": {
              "type": "string"
            },
            "escola": {
              "type": "string"
            },
            "email":{
              "type":"string"
            },
            "confirmacao":{
              "type":"integer"
            },
            "premium":{
              "type":"integer"
            },
            "validade":{
              "type":"string"
            },
            "socionum":{
              "type":"integer"
            },
            "projeto":{
              "type":"integer"
            }
          }    
        },
...
        "Escola":{
          "type": "object",
          "properties": {
            "id":{
              "type": "integer"
            },
            "nome":{
              "type":"string"
            },
            "localidade":{
              "type":"string"
            },
            "distrito":{
              "type":"string"
            },
            "pais":{
              "type":"string"
            },
            "cod":{
              "type":"string"
            }
          }
        }
    }   
}